requirements.txt
absl-py==1.4.0 ; python_version >= '3.6'
aiohttp==3.8.3
aiosignal==1.3.1 ; python_version >= '3.7'
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==22.2.0 ; python_version >= '3.6'
cachetools==5.2.1 ; python_version ~= '3.7'
certifi==2022.12.7 ; python_version >= '3.6'
charset-normalizer==2.1.1 ; python_full_version >= '3.6.0'
click==8.1.3
contourpy==1.0.7 ; python_version >= '3.8'
cycler==0.11.0 ; python_version >= '3.6'
docker-pycreds==0.4.0
fonttools==4.38.0 ; python_version >= '3.7'
frozenlist==1.3.3 ; python_version >= '3.7'
fsspec[http]==2023.1.0 ; python_version >= '3.7'
future==0.18.3 ; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gitdb==4.0.10 ; python_version >= '3.7'
gitpython==3.1.30 ; python_version >= '3.7'
google-auth==2.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-auth-oauthlib==0.4.6 ; python_version >= '3.6'
grpcio==1.51.1 ; python_version >= '3.7'
hydra-core==1.3.1
idna==3.4 ; python_version >= '3.5'
importlib-metadata==6.0.0 ; python_version < '3.10'
importlib-resources==5.10.2 ; python_version < '3.9'
kiwisolver==1.4.4 ; python_version >= '3.7'
markdown==3.4.1 ; python_version >= '3.7'
markupsafe==2.1.2 ; python_version >= '3.7'
matplotlib==3.5.3
multidict==6.0.4 ; python_version >= '3.7'
numpy==1.24.1 ; python_version >= '3.8'
oauthlib==3.2.2 ; python_version >= '3.6'
omegaconf==2.3.0 ; python_version >= '3.6'
packaging==23.0 ; python_version >= '3.7'
pathtools==0.1.2
pillow==9.4.0 ; python_version >= '3.7'
protobuf==3.20.3 ; python_version < '3.9' and sys_platform == 'linux'
psutil==5.9.4 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeprecate==0.3.1 ; python_version >= '3.6'
pyparsing==3.0.9 ; python_full_version >= '3.6.8'
python-dateutil==2.8.2 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pytorch-lightning==1.4.9
pyyaml==6.0 ; python_version >= '3.6'
requests==2.28.2 ; python_version >= '3.7' and python_version < '4'
requests-oauthlib==1.3.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
rsa==4.9 ; python_version >= '3.6'
scipy==1.7.3
sentry-sdk==1.13.0
setproctitle==1.3.2 ; python_version >= '3.7'
setuptools==66.1.1 ; python_version >= '3.7'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
smmap==5.0.0 ; python_version >= '3.6'
tensorboard==2.11.2 ; python_version >= '3.7'
tensorboard-data-server==0.6.1 ; python_version >= '3.6'
tensorboard-plugin-wit==1.8.1
torch==1.11
torchmetrics==0.11.0 ; python_version >= '3.7'
torchvision==0.12
tqdm==4.64.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.4.0 ; python_version >= '3.7'
urllib3==1.26.14 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wandb==0.13.9
werkzeug==2.2.2 ; python_version >= '3.7'
wheel==0.38.4 ; python_version >= '3.7'
yarl==1.8.2 ; python_version >= '3.7'
zipp==3.11.0 ; python_version < '3.10'